remove os sync action and opensearch write logic
