+++ app/commands.py
@@ -1,5 +1,8 @@
 import os
 import posixpath
+
+# ruff: noqa: F401
+# time isn't used here but patched in a test for retry
 import time
 import xml.etree.ElementTree as ET
 from datetime import date, datetime, timedelta, timezone
@@ -8,9 +11,9 @@ from urllib.parse import urlparse
 
 import click
 from flask import Blueprint
-from opensearchpy.exceptions import ConnectionTimeout, OpenSearchException
 from opensearchpy.helpers import scan
-from sqlalchemy.exc import OperationalError
+
+from tests.helpers.opensearch import index_datasets, refresh
 
 from .database import CatalogDBInterface, OpenSearchInterface
 from .models import (
@@ -93,212 +96,6 @@ def load_test_data(clear):
         interface.db.rollback()
         click.echo(f"Error loading test data: {e}")
         raise
-
-
-@search.cli.command("sync")
-@click.argument("dataset_id_or_slug", required=False)
-@click.option("--start-page", help="Number of page to start on", default=1)
-@click.option("--per_page", help="Number of datasets per page", default=100)
-@click.option(
-    "--recreate-index",
-    is_flag=True,
-    help="Delete and recreate index with new schema",
-    default=False,
-)
-def sync_opensearch(
-    dataset_id_or_slug: Optional[str] = None,
-    start_page: int = 1,
-    per_page: int = 100,
-    recreate_index: bool = False,
-):
-    """Sync datasets to the OpenSearch system.
-
-    Provide a DATASET_ID_OR_SLUG argument to reindex a single dataset without
-    touching the rest of the index.
-
-    Use --recreate-index flag when you've updated the schema (e.g., added keyword.raw field)
-    to delete the old index and create a new one with the updated mapping.
-
-    Retries added for when we may have multiple jobs running and causes the sync to break.
-    There are 3 exponential retries, the initial retry being 2 seconds.
-    """
-
-    # Retry configuration
-    max_retries = 3
-    retry_delay = 2.0
-    opensearch_errors = []
-
-    client = OpenSearchInterface.from_environment()
-    interface = CatalogDBInterface()
-
-    # get the count of datasets before new indexing
-    pre_os_dataset_count = client.count_all_datasets()
-
-    if dataset_id_or_slug:
-        interface = CatalogDBInterface()
-        if recreate_index:
-            raise click.ClickException(
-                "Cannot use --recreate-index when syncing a single dataset."
-            )
-
-        dataset = interface.get_dataset_by_id(dataset_id_or_slug)
-        if dataset is None:
-            dataset = interface.get_dataset_by_slug(dataset_id_or_slug)
-
-        if dataset is None:
-            raise click.ClickException(
-                f"Dataset '{dataset_id_or_slug}' was not found by id or slug."
-            )
-
-        click.echo(
-            f"Indexing dataset {dataset.id} (slug: {dataset.slug}) into OpenSearch..."
-        )
-        succeeded, failed, errors = client.index_datasets([dataset])
-
-        if failed:
-            raise click.ClickException(
-                f"Failed to index dataset {dataset.id}; see logs for details."
-            )
-
-        click.echo("Dataset indexed successfully.")
-        return
-
-    # empty the index and then refill it
-    # THIS WILL CAUSE INCONSISTENT SEARCH RESULTS DURING THE PROCESS
-
-    if recreate_index:
-        click.echo("Deleting entire index to recreate with new schema...")
-        try:
-            client.client.indices.delete(index=client.INDEX_NAME)
-            click.echo("Index deleted")
-        except Exception as e:
-            click.echo(f"Could not delete index (may not exist): {e}")
-
-        # Recreate with new schema
-        click.echo("Creating index with new schema...")
-        client._ensure_index()
-        click.echo("Index created with updated mapping")
-
-        # Verify the new mapping
-        mapping = client.client.indices.get_mapping(index=client.INDEX_NAME)
-        keyword_mapping = mapping[client.INDEX_NAME]["mappings"]["properties"].get(
-            "keyword", {}
-        )
-        has_raw = "fields" in keyword_mapping and "raw" in keyword_mapping["fields"]
-        if has_raw:
-            click.echo("Verified: keyword.raw field exists in new mapping")
-        else:
-            click.echo(
-                "Warning: keyword.raw field not found in mapping - aggregations may not work"
-            )
-    else:
-        click.echo("Emptying dataset index (keeping existing schema)...")
-        client.delete_all_datasets()
-
-    click.echo("Indexing...")
-
-    # do our own pagination of the dataset query before calling into the
-    # index_datasets method
-    total_pages = Dataset.query.paginate(per_page=per_page).pages
-    click.echo(f"Indexing {total_pages} pages of datasets...")
-
-    try:
-        # page numbers are 1-indexed
-        for i in range(start_page, total_pages + 1):
-            retry_count = 0
-            last_exception = None
-
-            while retry_count <= max_retries:
-                try:
-                    # Get the paginated dataset query
-                    paginated_datasets = Dataset.query.paginate(
-                        page=i, per_page=per_page
-                    )
-
-                    # Index the datasets
-                    succeeded, failed, errors = client.index_datasets(
-                        paginated_datasets, refresh_after=False
-                    )
-                    # add errors to render later
-                    opensearch_errors.extend(errors)
-                    # Success - break out of retry loop
-                    click.echo(
-                        f"Indexed page {i}/{total_pages} with {succeeded} successes and {failed} errors."
-                    )
-                    break
-
-                except (OpenSearchException, ConnectionTimeout, OperationalError) as e:
-                    last_exception = e
-                    retry_count += 1
-
-                    error_type = type(e).__name__
-
-                    # Check if this is a PostgreSQL serialization failure
-                    # Safely convert exception to string
-                    try:
-                        error_str = str(e)
-                    except Exception:
-                        error_str = repr(e)
-
-                    is_serialization_error = (
-                        isinstance(e, OperationalError)
-                        and "conflict with recovery" in error_str
-                    )
-
-                    if retry_count <= max_retries:
-                        # Calculate exponential backoff delay
-                        wait_time = retry_delay * (2 ** (retry_count - 1))
-
-                        click.echo(
-                            f"Page {i}/{total_pages}: {error_type} - "
-                            f"{'Database serialization conflict' if is_serialization_error else 'Error'} "
-                            f"(attempt {retry_count}/{max_retries + 1}). "
-                            f"Retrying in {wait_time:.1f} seconds..."
-                        )
-                        time.sleep(wait_time)
-                        continue
-                    else:
-                        # Max retries exceeded
-                        try:
-                            error_msg = str(last_exception)
-                        except Exception:
-                            error_msg = repr(last_exception)
-                        click.echo(
-                            f"Page {i}/{total_pages}: Failed after {max_retries + 1} attempts. "
-                            f"Last error: {error_type} - {error_msg[:200] if error_msg else 'Unknown error'}"
-                        )
-                        # Exit with error code
-                        raise click.ClickException(
-                            f"Sync failed after {max_retries + 1} attempts"
-                        )
-
-        click.echo("Refreshing index...")
-        client._refresh()
-        click.echo("Sync was successful")
-    except click.ClickException:
-        # Re-raise Click exceptions (these exit cleanly with proper exit code)
-        raise
-    except Exception as e:
-        # Catch any other unexpected errors
-        click.echo(f"Unexpected error during sync: {type(e).__name__}")
-        raise click.ClickException(f"Sync failed: {type(e).__name__}")
-
-    click.echo("=" * 20 + "STATS" + "=" * 20)
-    click.echo(f"Total Datasets in Database: {interface.total_datasets()}")
-    click.echo(f"Total Datasets in Index Before Sync: {pre_os_dataset_count}")
-    click.echo(f"Total Datasets in Index After Sync: {client.count_all_datasets()}")
-    click.echo(f"Recreate Index: {recreate_index}")
-    click.echo(f"Total Errors: {len(opensearch_errors)}")
-    if opensearch_errors:
-        click.echo("=" * 20 + "ERRORS" + "=" * 20)
-        for opensearch_error in opensearch_errors:
-            click.echo(
-                f"Dataset ID: {opensearch_error.get("dataset_id")}, "
-                f"Status Code: {opensearch_error.get("status_code")}, "
-                f"Error Type: {opensearch_error.get("error_type")}, "
-                f"Error Reason: {opensearch_error.get("error_reason")}, "
-                f"Caused By: {opensearch_error.get("caused_by")}"
-            )
 
 
 @search.cli.command("compare")
@@ -393,8 +190,8 @@ def compare_opensearch(sample_size: int, update: bool, force_update: bool):
                 )
 
             if datasets:
-                succeeded, failed, errors = client.index_datasets(
-                    datasets, refresh_after=False
+                succeeded, failed, errors = index_datasets(
+                    client, datasets, refresh_after=False
                 )
                 total_indexed += succeeded
                 if failed:
@@ -470,8 +267,7 @@ def compare_opensearch(sample_size: int, update: bool, force_update: bool):
         click.echo("Example extra IDs: none")
 
     click.echo(
-        "Updated in OpenSearch (last_harvested_date differs): "
-        f"{len(updated_details)}"
+        f"Updated in OpenSearch (last_harvested_date differs): {len(updated_details)}"
     )
     if updated_details:
         sample_entries = [
@@ -543,7 +339,7 @@ def compare_opensearch(sample_size: int, update: bool, force_update: bool):
 
     if missing or extra or updated_ids or force_reindex_ids:
         click.echo("Refreshing OpenSearch index…")
-        client._refresh()
+        refresh(client)
         click.echo("Done.")
     else:
         click.echo("Nothing to update; datasets and index are already in sync.")
